requirements.txt
flask
flask-cors
pymongo
pandas
scikit-learn
python-dotenv
dnspython